Combining TestHarness and WindowsService projects. Issues/15. * Moved Topshelf logic into the TestHarness. TODO: Renaming TestHarness project/namespaces should probably follow. * Removes the old WindowsService and housing solution folder. * Adds Topshelf package to source control, and removed packages from .gitignore; Reduces the number of sites that have   to be online in order to get and build source successfully. (We've been bitten by offline/rogue nugets before.) * If run without arguments, TestHarness runs in the normal interactive mode, else runs in Topshelf/service mode. * Simplified that assembly name to "WheelMUD.exe"
