+++ .gitignore
@@ -132,14 +132,7 @@ publish/
 *.pubxml
 *.publishproj
 
-# NuGet Packages
-*.nupkg
-# The packages folder can be ignored because of Package Restore
-**/packages/*
-# except build/, which is used as an MSBuild target.
-!**/packages/build/
-# If using the old MSBuild-Integrated Package Restore, uncomment this:
-#!**/packages/repositories.config
+# NuGet Packages are intentionally NOT ignored here.
 
 # Windows Azure Build Output
 csx/
