--- src/ServerAdministration/WindowsService/App.config
@@ -1,96 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
-  </configSections>
-  <appSettings>
-    <add key="DefaultConnectionStringName" value="WheelMUDSQLite" />
-    <add key="ClientSettingsProvider.ServiceUri" value="" />
-  </appSettings>
-  <connectionStrings configSource="Files\connectionstrings.config" />
-  <log4net>
-    <appender name="AdoNetAppender" type="log4net.Appender.AdoNetAppender">
-      <bufferSize value="100" />
-      <connectionType value="Finisar.SQLite.SQLiteConnection, SQLite.NET, Version=0.21.1869.3794, Culture=neutral, PublicKeyToken=c273bd375e695f9c" />
-      <connectionString value="Data Source=c:\\inetpub\\wwwroot\\logs\\log4net.db;Version=3;" />
-      <commandText value="INSERT INTO Log (Date, Level, Logger, Message) VALUES (@Date, @Level, @Logger, @Message)" />
-      <parameter>
-        <parameterName value="@Date" />
-        <dbType value="DateTime" />
-        <layout type="log4net.Layout.RawTimeStampLayout" />
-      </parameter>
-      <parameter>
-        <parameterName value="@Level" />
-        <dbType value="String" />
-        <layout type="log4net.Layout.PatternLayout">
-          <conversionPattern value="%level" />
-        </layout>
-      </parameter>
-      <parameter>
-        <parameterName value="@Logger" />
-        <dbType value="String" />
-        <layout type="log4net.Layout.PatternLayout">
-          <conversionPattern value="%logger" />
-        </layout>
-      </parameter>
-      <parameter>
-        <parameterName value="@Message" />
-        <dbType value="String" />
-        <layout type="log4net.Layout.PatternLayout">
-          <conversionPattern value="%message" />
-        </layout>
-      </parameter>
-    </appender>
-    <appender name="EventLogAppender" type="log4net.Appender.EventLogAppender">
-      <applicationName value="WheelMUD" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%property{NDC}] - %message%newline" />
-      </layout>
-    </appender>
-    <appender name="RollingLogFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file value="logfile.txt" />
-      <appendToFile value="false" />
-      <rollingStyle value="Size" />
-      <maxSizeRollBackups value="-1" />
-      <maximumFileSize value="5MB" />
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [%thread] %-5level %logger [%property{NDC}] - %message%newline" />
-      </layout>
-    </appender>
-    <root>
-      <level value="INFO" />
-      <appender-ref ref="RollingLogFileAppender" />
-    </root>
-  </log4net>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="NHibernate" publicKeyToken="AA95F207798DFDB4" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-2.1.0.4000" newVersion="2.1.2.4000" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Raven.Abstractions" publicKeyToken="37f41c7f99471593" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-2.0.3.0" newVersion="2.0.3.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Raven.Client.Lightweight" publicKeyToken="37f41c7f99471593" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-2.0.3.0" newVersion="2.0.3.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-  <startup useLegacyV2RuntimeActivationPolicy="true">
-    <supportedRuntime version="v4.0" />
-  </startup>
-  <system.web>
-    <membership defaultProvider="ClientAuthenticationMembershipProvider">
-      <providers>
-        <add name="ClientAuthenticationMembershipProvider" type="System.Web.ClientServices.Providers.ClientFormsAuthenticationMembershipProvider, System.Web.Extensions, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" serviceUri="" />
-      </providers>
-    </membership>
-    <roleManager defaultProvider="ClientRoleProvider" enabled="true">
-      <providers>
-        <add name="ClientRoleProvider" type="System.Web.ClientServices.Providers.ClientRoleProvider, System.Web.Extensions, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" serviceUri="" cacheTimeout="86400" />
-      </providers>
-    </roleManager>
-  </system.web>
-</configuration>
