--- src/ServerAdministration/WindowsService/Properties/AssemblyInfo.cs
@@ -1,26 +1,0 @@
-//-----------------------------------------------------------------------------
-// <copyright file="AssemblyInfo.cs" company="WheelMUD Development Team">
-//   Copyright (c) WheelMUD Development Team.  See LICENSE.txt.  This file is 
-//   subject to the Microsoft Public License.  All other rights reserved.
-// </copyright>
-// <summary>
-//   Assembly information for the WheelMUD.Utilities project.
-// </summary>
-//-----------------------------------------------------------------------------
-
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("WheelMUD.WindowsService")]
-[assembly: AssemblyDescription("A Windows Service to run the WheelMUD Server.")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("39FB9B32-0341-499c-A1AE-7C55CCBC1E2F")]
