--- src/ServerAdministration/WindowsService/mud.config
@@ -1,51 +1,0 @@
-<?xml version="1.0"?>
-<!-- 
-    Note: As an alternative to hand editing this file you can use the 
-    web admin tool to configure settings for your application. Use
-    the Website->Asp.Net Configuration option in Visual Studio.
-    A full list of settings and comments can be found in 
-    machine.config.comments usually located in 
-    \Windows\Microsoft.Net\Framework\v2.x\Config 
--->
-<configuration>
-    
-    <configSections>
-        <section name="EngineAttributes" type="System.Configuration.NameValueSectionHandler" />
-        <section name="UniverseAttributes" type="System.Configuration.NameValueSectionHandler" />
-        <section name="Currency" type="System.Configuration.NameValueSectionHandler" />
-        <section name="RuleSetAttributes" type="System.Configuration.NameValueSectionHandler" />
-        <section name="Templates" type="System.Configuration.NameValueSectionHandler" />
-        <section name="FTP" type="System.Configuration.NameValueSectionHandler" />
-    </configSections>
-    
-    <EngineAttributes>
-        <add key="name" value="WheelMUD.net" />
-        <add key="website" value="www.wheelmud.net" />
-        <add key="port" value="4000" />        
-    </EngineAttributes>
-
-    <UniverseAttributes>
-        <add key="defaultroom" value="1"/>        
-    </UniverseAttributes>
-
-    <Currency>
-      <add key="singular" value="gold coin" />
-      <add key="plural" value="gold coins" />
-    </Currency>
-
-    <RuleSetAttributes>
-      <add key="masterfile" value="WarriorRogueMageMaster.xml"/>
-      <add key="currentruleset" value="Warrior, Rogue, and Mage"/>
-    </RuleSetAttributes>
-
-    <Templates>
-      <add key="helpviewtemplate" value="StandardHelpTopicViewTemplate.vm" />
-      <add key="roomviewtemplate" value="StandardRoomViewTemplate.vm" />
-       <add key="entityviewtemplate" value="StandardEntityViewTemplate.vm" />
-    </Templates>
-
-    <FTP>
-      <add key="RootFolder" value="%FTPRUNDIR%\Files"/>
-    </FTP>
-    
-</configuration>
