+++ src/packages/repositories.config
@@ -7,5 +7,6 @@
   <repository path="..\Data\packages.config" />
   <repository path="..\Server\packages.config" />
   <repository path="..\ServerAdministration\WindowsService\packages.config" />
+  <repository path="..\TestHarness\packages.config" />
   <repository path="..\WarriorRogueMage\packages.config" />
 </repositories>
